+++ Sources/HIGThemesContract/HIGTheme.swift
@@ -35,6 +35,7 @@ public protocol HIGTheme: Sendable {
     var tooltip: any HIGTooltipTokens { get }
     var popover: any HIGPopoverTokens { get }
     var drawer: any HIGDrawerTokens { get }
+    var networkProgressBar: any HIGNetworkProgressBarTokens { get }
     var progress: any HIGProgressTokens { get }
     var alert: any HIGAlertTokens { get }
     var toast: any HIGToastTokens { get }